Added a component for reusability
