+++ apps/web/components/real.ts
@@ -76,3 +76,4 @@ export * from './NewLinks/NewLinks'
 export * from './NewsItems/NewsItems'
 export * from './GridItems/GridItems'
 export * from './LifeEventsSection/LifeEventsSection'
+export * from './UkraineChatPanel/UkraineChatPanel'
